Added support for loading geometry shaders, added optimization to use vertex array objects automatically when they are supported, added support for uniform buffer objects, reduced amount of redundant OpenGL API calls through more aggressive state caching, added methods to retrieve buffer object and shader object names, got rid of pre-transformed vertex cache (it didn't really make much sense), moved light component color computations out of fragment shader and into Light class, combined light attenuation values into a single vec4, moved static variables out of Light class and into its translation unit, fixed a few type conversion warnings, added identifier to Config.hpp to allow checking for existence of 3DEE.

+++ src/SFML/Graphics/SphericalPolyhedron.cpp
@@ -57,7 +57,7 @@ unsigned int SphericalPolyhedron::getFaceCount() const
     if (m_geometry.empty())
         construct();
 
-    return m_geometry.size() / 3;
+    return static_cast<unsigned int>(m_geometry.size()) / 3;
 }
